Temporarily switched the Player to a DynamicObject
Because the frame rate variable was an int, the delta time calculation ended up being rounded to 0.

+++ Kernmodule GDEV Game/Kernmodule GDEV Game.cpp
@@ -6,7 +6,7 @@ int main()
 {
 	int screenWidth = 512;
 	int screenHeight = 512;
-	int frameRate = 60;
+	float frameRate = 60;
 	float deltaTime = 1 / frameRate;
 
 	sf::RenderWindow window(sf::VideoMode(screenWidth, screenHeight), "I have no idea what I'm doing.");
